+++ python/manual.txt
